+++ Point.h
@@ -7,23 +7,23 @@ class Point
 public:
 
 	Point();
-	void setPolar(float angle, float distance);
+	void setPolar(double angle, double distance);
 
-	float getAngle() const;
-	float getDistance() const;
+	double getAngle() const;
+	double getDistance() const;
 
-	float getX(int i = 0) const;
-	float getY(int i = 0) const;
-	//float getX_S() const;
-	//float getY_S() const;
+	double getX(int i = 0) const;
+	double getY(int i = 0) const;
+	//double getX_S() const;
+	//double getY_S() const;
 	
 private:
 
-	float _angle;
-	float _distance;
-	float _coorX;
-	float _coorY;
-	float _coorX_S;
-	float _coorY_S;
+	double _angle;
+	double _distance;
+	double _coorX;
+	double _coorY;
+	double _coorX_S;
+	double _coorY_S;
 
 };
